use a single DomStorageMock definition
